chore: Moved data mapper to another package
